requirements.txt
asgiref==3.4.1
Django>=3.2.15
pytz==2021.3
sqlparse>=0.4.2

django-background-tasks==1.2.5
dj-database-url==0.5.0
gunicorn==22.0.0
psycopg2-binary==2.8.6
whitenoise==5.2.0

numpy>=1.22
pandas==1.3.4
scikit-learn==1.0.2
mne>=1.3.1
scipy>=1.10.0
joblib>=1.2.0
celery==5.2.6
django-redis==5.0.0
django-celery-beat==2.2.1
django-celery-results>=2.4.0
django-widget-tweaks>=1.4.12
redis>=4.4.4
django-storages[azure]==1.13.1
azure-storage-blob==12.13.1
tqdm>=4.64.1
aiohttp>=3.8.4
python-dotenv=1.0.1